Move Name to domain/shared package
