+++ src/test/java/com/alvarx4z/woja/domain/CharacterTest.java
@@ -1,9 +1,9 @@
 package com.alvarx4z.woja.domain;
 
+import com.alvarx4z.woja.domain.shared.Name;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
-import static com.alvarx4z.woja.helper.CharacterHelper.aCharacter;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.junit.jupiter.api.Assertions.assertThrows;
 
@@ -12,7 +12,7 @@ final class CharacterTest {
     @Test
     @DisplayName("Should instantiate correctly a Character")
     void givenCharacter_whenInstantiating_thenInstantiatedCorrectly() {
-        Character character = aCharacter();
+        Character character = new Character(Name.of("Vol'jin"));
 
         assertThat(character).isInstanceOf(Character.class);
         assertThat(character.getName()).isInstanceOf(Name.class);
